+++ src/ngraph/runtime/CMakeLists.txt
@@ -16,7 +16,9 @@
 
 add_subdirectory(interpreter)
 
-add_subdirectory(hybrid)
+if (NGRAPH_HYBRID_ENABLE)
+    add_subdirectory(hybrid)
+endif()
 
 if (NGRAPH_CPU_ENABLE)
     add_subdirectory(cpu)
