+++ test/hybrid_utils.cpp
@@ -154,42 +154,3 @@ bool TestBackend::call(shared_ptr<Function> func,
     }
     return rc;
 }
-
-BackendWrapper::BackendWrapper(const string& backend_name,
-                               const set<string>& supported_ops,
-                               const string& name)
-    : m_backend{runtime::Backend::create(backend_name)}
-    , m_supported_ops{supported_ops}
-    , m_name{name}
-{
-}
-
-shared_ptr<runtime::Tensor> BackendWrapper::create_tensor(const element::Type& element_type,
-                                                          const Shape& shape)
-{
-    return m_backend->create_tensor(element_type, shape);
-}
-
-shared_ptr<runtime::Tensor> BackendWrapper::create_tensor(const element::Type& element_type,
-                                                          const Shape& shape,
-                                                          void* memory_pointer)
-{
-    return m_backend->create_tensor(element_type, shape, memory_pointer);
-}
-
-bool BackendWrapper::compile(shared_ptr<Function> func)
-{
-    return m_backend->compile(func);
-}
-
-bool BackendWrapper::call(shared_ptr<Function> func,
-                          const vector<shared_ptr<runtime::Tensor>>& outputs,
-                          const vector<shared_ptr<runtime::Tensor>>& inputs)
-{
-    return m_backend->call(func, outputs, inputs);
-}
-
-bool BackendWrapper::is_supported(const Node& node) const
-{
-    return m_supported_ops.find(node.description()) != m_supported_ops.end();
-}
